server: use HTTP 400 where there is a client error
We often send a 500 error where it should be 400 or 404.
It is important to get these right.
A client will assume it did nothing wrong when a 500 is received
and perhaps automatically retry the request.

This is a conservative change:
there are many more places that should be changed from 500.
A helper function jsonRespondError is added